Add explicit RedshiftDatabaseTasks and correct Hash/kwargs confusion
This adds explicit (copy-pasted from the PostgreSQLDatabaseTasks in
Rails) DatabaseTasks for Redshift. The only change is to make the public
schema connection to a database named `redshift` not `postgres`.

It also partially-reverts

https://github.com/pennylane-hq/activerecord-adapter-redshift/commit/1ce6bdbcba7ade96960886895e7ff1f85d5263c0

to be in-line with the Postgres adapter, where some Hash options still
remain and haven't been converted to kwargs.

This avoids the error:
```
~/.rbenv/versions/3.3.3/lib/ruby/gems/3.3.0/gems/activerecord7-redshift-adapter-pennylane-1.0.4/lib/active_record/connection_adapters/redshift_7_1/schema_statements.rb:40:in `create_database': wrong number of arguments (given 2, expected 1) (ArgumentError)
```